[Silver III] Title: 주유소, Time: 300 ms, Memory: 40484 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -22,7 +22,7 @@ public class Main {
         }
 
         for(int i = 0; i<N-1; i++){
-            answer += prices[i] * dist[i];
+            answer += (long) prices[i] * dist[i];
         }
 
         System.out.println(answer);
